Removed time step restrictions on Imaginary integral.
Restricting the time step does not work so well for tiny nu values. It
seems better to stick with the older method of restricting the
integration range when nu became small. Small here means when the arctan
starts to look like a step function. Of course, this is also a relative
statement, and I am not sure how to make this less empirical.